# Modify test to improve performance