--- a/dev/com.ibm.ws.request.timing_fat/fat/src/com/ibm/ws/request/timing/fat/HungRequestTiming.java
+++ b/dev/com.ibm.ws.request.timing_fat/fat/src/com/ibm/ws/request/timing/fat/HungRequestTiming.java
@@ -91,11 +91,11 @@ public class HungRequestTiming {
 
         assertNotNull("The server configuration was successfully updated message was not found!", srvConfigCompletedMsg);
 
-        CommonTasks.writeLogMsg(Level.INFO, "************ Thread 1 - Runs 3s  ************");
-        CommonTasks.writeLogMsg(Level.INFO, "************ Thread 2 - Runs 5s ************");
+        CommonTasks.writeLogMsg(Level.INFO, "************ Thread 1 - Runs 4s  ************");
+        CommonTasks.writeLogMsg(Level.INFO, "************ Thread 2 - Runs 6s ************");
 
-        HungRequestThread request1 = new HungRequestThread(3000);
-        HungRequestThread request2 = new HungRequestThread(5000);
+        HungRequestThread request1 = new HungRequestThread(4000);
+        HungRequestThread request2 = new HungRequestThread(6000);
 
         request1.start();
         request2.start();