--- a/dev/com.ibm.ws.request.timing_fat/publish/files/server_hungRequestThreshold2.xml
+++ b/dev/com.ibm.ws.request.timing_fat/publish/files/server_hungRequestThreshold2.xml
@@ -1,0 +1,16 @@
+<server description="new server">
+
+  <!-- Enable features -->
+    <featureManager>
+        <feature>jsp-2.2</feature>
+      <feature>requestTiming-1.0</feature>
+    </featureManager>
+  
+	<httpEndpoint id="defaultHttpEndpoint" host="*" />
+
+	<include location="../fatTestPorts.xml"/>
+  	
+   <requestTiming slowRequestThreshold="0"/>
+   <requestTiming hungRequestThreshold = "2s" />
+  
+</server>