--- a/dev/com.ibm.ws.request.timing_fat/publish/files/server_sampleRate0.xml
+++ b/dev/com.ibm.ws.request.timing_fat/publish/files/server_sampleRate0.xml
@@ -27,5 +27,6 @@
  	<include location="../fatTestPorts.xml"/>
    
 	<requestTiming sampleRate = "0" />
+	<requestTiming slowRequestThreshold="2s" />
   
 </server>