… dev/com.ibm.ws.request.timing_fat/fat/src/com/ibm/ws/request/timing/fat/TimingRequestTiming.java
… dev/com.ibm.ws.request.timing_fat/publish/files/server_slowRequestThreshold1.xml