DO NOT MERGE! Work still in progress.  Status ---> In for final review

The latest build is located here https://wasrtc.hursley.ibm.com:9443/jazz/resource/itemOid/com.ibm.team.build.BuildResult/_2aVDgMS2Eeq1S7zZzzGuaA
The build completed successfully without any failures or errors.

The following are the run times **before** the changes...
**Lite  --- Total approx.  26.5min**
HungRequestTiming 2m 26s
SlowRequestTiming 11m 52s
TimingRequestTiming 12m 07s
**Full  --- Total approx.  96.5min**
HungRequestTiming 16m 40s
SlowRequestTiming 46m 34s
TimingRequestTiming 33m 30s

The following are the run times **after** the changes...
**Lite  --- Total approx.  15min**
HungRequestTiming 1m 51s
SlowRequestTiming 6m 09s
TimingRequestTiming 6m 57s
**Full  --- Total approx.  59min**
HungRequestTiming 15m 07s
SlowRequestTiming 23m 50s
TimingRequestTiming 19m 42s

The following is a list of changes...
**List of changes in regards to increasing the performance and runtime of the tests**
1. Removed a lot of unnecessary server.setMarkToEndOfLog calls.
2. Removed unnecessary sleeps throughout the test cases and test applications.
3. Shorten the sleeps and hung/slow request thresholds where possible  (EX/ Instead of waiting 10 sec for a slow request we only have to wait 3+sec if we set the threshold to 3sec).
4. Removed many unnecessary server configuration changes/updates.
5. Improved the TestJDBC test applications to execute updates without having to wait for a hardcoded 5secs for each update and also removed unnecessary thread sleeps. This saves about 30secs per request. If a test case makes two requests then we save 1min of unnecessary waits.

**List of bug fixes and other fixes**
1. Corrected many log messages that contained wrong information
2. Corrected many comments in the code which contained wrong information
3. Fixed a rare synchronization bug in one of the test cases for HungRequestTiming where we can't find a message in the log since we did not give enough time for it to generate